Refactor contact hands settings into its own scriptable object
Attempt to get first load working
